Version 2.7

+++ 42 Cursus/Libft/ft_lstsize.c
@@ -6,7 +6,7 @@
 /*   By: sde-carv <sde-carv@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/17 17:10:01 by sde-carv          #+#    #+#             */
-/*   Updated: 2024/05/18 09:46:12 by sde-carv         ###   ########.fr       */
+/*   Updated: 2024/05/18 18:50:50 by sde-carv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_lstsize(t_list *lst)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (lst)
